Mes 3649 restructuring folder for wrtc page (#816)
* mes-3649-restructuredFolderForWRTCPage

* mes-3649-Renamed files,classes,selectors and page name constant etc

* mes-3649-finishedRebase

* mes-3649-renamedSpecTestToContainCatB
