+++ src/pages/page-names.constants.ts
@@ -13,13 +13,13 @@ export const VIEW_TEST_RESULT_PAGE = 'ViewTestResultPage';
 export const CAT_B = {
   BACK_TO_OFFICE_PAGE: 'BackToOfficePage',
   COMMUNICATION_PAGE: 'CommunicationCatBPage',
-  DEBRIEF_PAGE:'DebriefPage',
+  DEBRIEF_PAGE: 'DebriefPage',
   HEALTH_DECLARATION_PAGE: 'HealthDeclarationPage',
   OFFICE_PAGE: 'OfficePage',
   PASS_FINALISATION_PAGE: 'PassFinalisationPage',
-  TEST_REPORT_PAGE:'TestReportPage',
+  TEST_REPORT_PAGE: 'TestReportPage',
   WAITING_ROOM_PAGE: 'WaitingRoomCatBPage',
-  WAITING_ROOM_TO_CAR_PAGE: 'WaitingRoomToCarPage',
+  WAITING_ROOM_TO_CAR_PAGE: 'WaitingRoomToCarCatBPage',
   POST_DEBRIEF_HOLDING_PAGE: 'PostDebriefHoldingPage',
   REKEY_REASON_PAGE: 'RekeyReasonPage',
   REKEY_UPLOAD_OUTCOME_PAGE: 'RekeyUploadOutcomePage',
@@ -29,11 +29,11 @@ export const CAT_B = {
 export const CAT_BE = {
   BACK_TO_OFFICE_PAGE: 'BackToOfficePage',
   COMMUNICATION_PAGE: 'CommunicationCatBPage',
-  DEBRIEF_PAGE:'DebriefPage',
+  DEBRIEF_PAGE: 'DebriefPage',
   HEALTH_DECLARATION_PAGE: 'HealthDeclarationPage',
   OFFICE_PAGE: 'OfficePage',
   PASS_FINALISATION_PAGE: 'PassFinalisationPage',
-  TEST_REPORT_PAGE:'TestReportPage',
+  TEST_REPORT_PAGE: 'TestReportPage',
   WAITING_ROOM_PAGE: 'WaitingRoomCatPage',
   WAITING_ROOM_TO_CAR_PAGE: 'WaitingRoomToCarPage',
   POST_DEBRIEF_HOLDING_PAGE: 'PostDebriefHoldingPage',
