+++ src/pages/waiting-room-to-car/cat-b/waiting-room-to-car.cat-b.page.html
@@ -1,0 +1,62 @@
+<ion-header>
+  <practice-mode-banner *ngIf="isPracticeMode"></practice-mode-banner>
+  <ion-navbar hideBackButton>
+    <ion-title>{{pageState.candidateName$ | async}}</ion-title>
+    <ion-buttons end>
+      <end-test-link></end-test-link>
+    </ion-buttons>
+  </ion-navbar>
+</ion-header>
+
+<ion-content>
+  <div>
+    <form [formGroup]="form" (ngSubmit)="onSubmit()">
+      <ion-grid>
+
+        <eyesight-test [formGroup]="form" [eyesightPassRadioChecked]="(pageState.eyesightTestComplete$ | async) && !(pageState.eyesightTestFailed$ | async)"
+          [eyesightFailRadioChecked]="pageState.eyesightTestFailed$ | async" (eyesightTestResultChange)="eyesightTestResultChanged($event)"></eyesight-test>
+
+        <ion-row *ngIf="pageState.eyesightTestFailed$ | async">
+          <eyesight-failure-confirmation [cancelFn]="eyesightFailCancelled"></eyesight-failure-confirmation>
+        </ion-row>
+
+        <ion-row no-padding [hidden]="pageState.eyesightTestFailed$ | async" id="post-eyesight-form-content">
+          <ion-col no-padding>
+            <tell-me-question-card [formGroup]="form" [tellMeQuestion]="pageState.tellMeQuestion$ | async"
+              [tellMeQuestions]="tellMeQuestions" [tellMeQuestionOutcome]="pageState.tellMeQuestionOutcome$ | async"
+              [tellMeQuestionSelected]="pageState.tellMeQuestionSelected$ | async" (tellMeQuestionChange)="tellMeQuestionChanged($event)"
+              (tellMeQuestionOutcomeChange)="tellMeQuestionOutcomeChanged($event)">
+            </tell-me-question-card>
+
+            <vehicle-registration [formGroup]="form" [vehicleRegistration]="pageState.registrationNumber$ | async"
+              (vehicleRegistrationChange)="vehicleRegistrationChanged($event)"></vehicle-registration>
+
+            <transmission [formGroup]="form" [transmission]="pageState.transmission$ | async" (transmissionChange)="transmissionChanged($event)"></transmission>
+
+            <instructor-registration [formGroup]="form" [instructorRegistration]="pageState.instructorRegistrationNumber$ | async"
+              (instructorRegistrationChange)="instructorRegistrationChanged($event)"></instructor-registration>
+
+            <accompaniment-card [formGroup]="form" [instructorAccompaniment]="pageState.instructorAccompaniment$ | async"
+              [supervisorAccompaniment]="pageState.supervisorAccompaniment$ | async" [otherAccompaniment]="pageState.otherAccompaniment$ | async"
+              [interpreterAccompaniment]="pageState.interpreterAccompaniment$ | async" (instructorAccompanimentChange)="instructorAccompanimentToggled()"
+              (supervisorAccompanimentChange)="supervisorAccompanimentToggled()" (otherAccompanimentChange)="otherAccompanimentToggled()"
+              (interpreterAccompanimentChange)="interpreterAccompanimentToggled()">
+            </accompaniment-card>
+
+            <vehicle-details-card [formGroup]="form" [schoolVehicleDetails]="pageState.schoolCar$ | async"
+              [dualVehicleDetails]="pageState.dualControls$ | async" (schoolVehicleDetailsChange)="schoolCarToggled()"
+              (dualVehicleDetailsChange)="dualControlsToggled()"></vehicle-details-card>
+          </ion-col>
+        </ion-row>
+
+      </ion-grid>
+    </form>
+  </div>
+</ion-content>
+<ion-footer [hidden]="pageState.eyesightTestFailed$ | async">
+  <ion-row class="mes-full-width-card box-shadow">
+    <button type="submit" class="mes-primary-button" id="continue-to-test-report-button" ion-button (click)="onSubmit()">
+      <h3>Continue to test report</h3>
+    </button>
+  </ion-row>
+</ion-footer>
