+++ src/pages/waiting-room-to-car/cat-b/waiting-room-to-car.cat-b.page.scss
@@ -1,4 +1,4 @@
-page-waiting-room-to-car {
+waiting-room-to-car-cat-b-page {
   ion-content.content-ios {
     background-color: map-get($colors, "mes-white");
   }
