+++ src/pages/waiting-room-to-car/waiting-room-to-car.module.ts
@@ -1,20 +1,9 @@
 import { NgModule } from '@angular/core';
-import { IonicPageModule } from 'ionic-angular';
-import { WaitingRoomToCarPage } from './waiting-room-to-car';
-import { EffectsModule } from '@ngrx/effects';
-import { WaitingRoomToCarAnalyticsEffects } from '../waiting-room-to-car/waiting-room-to-car.analytics.effects';
-import { ComponentsModule } from '../../components/common/common-components.module';
-import { WaitingRoomToCarComponentsModule } from './components/waiting-room-to-car.components.module';
+import { WaitingRoomToCarCatBPageModule } from './cat-b/waiting-room-to-car.cat-b.page.module';
 
 @NgModule({
-  declarations: [
-    WaitingRoomToCarPage,
-  ],
   imports: [
-    IonicPageModule.forChild(WaitingRoomToCarPage),
-    EffectsModule.forFeature([WaitingRoomToCarAnalyticsEffects]),
-    ComponentsModule,
-    WaitingRoomToCarComponentsModule,
+    WaitingRoomToCarCatBPageModule,
   ],
 })
 export class WaitingRoomToCarPageModule { }
